+++ .eslintrc.js
@@ -9,10 +9,11 @@ module.exports = {
     '@vue/typescript'
   ],
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-console': 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-unused-vars': 'off',
     'vue/no-unused-vars': 'off',
+    'vue/multi-word-component-names': 'off',
   },
   parserOptions: {
     parser: '@typescript-eslint/parser'