+++ src/App.vue
@@ -47,8 +47,9 @@
 </template>
 
 <script setup lang="ts">
-import {ref, computed, onMounted, reactive} from "vue";
-import { type Vue } from 'vue-property-decorator';
+/* eslint-disable */
+import type Vue from "vue";
+import {ref, onMounted, reactive} from "vue";
 const PipingUI = () => import('@/components/PipingUI.vue');
 const MenuContent = () => import('@/components/MenuContent.vue');
 const Licenses = () => import("@/components/Licenses.vue");