+++ src/components/DarkThemeSwitch.vue
@@ -3,19 +3,22 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
+import Vue from 'vue';
 import {enableDarkTheme} from "@/enable-dark-theme";
 
-@Component
-export default class DarkThemeSwitch extends Vue {
-  get enableDarkTheme() {
-    return enableDarkTheme.value;
+export default Vue.extend({
+  computed: {
+    enableDarkTheme: {
+      get () {
+        return enableDarkTheme.value;
+      },
+      set (v: boolean) {
+        enableDarkTheme.value = v;
+        // Enable dark theme
+        // NOTE: `this.$vuetify.theme.dark` is not available in Composition API
+        this.$vuetify.theme.dark = v;
+      },
+    },
   }
-  set enableDarkTheme(v: boolean) {
-    enableDarkTheme.value = v;
-    // Enable dark theme
-    // NOTE: `this.$vuetify.theme.dark` is not available in Composition API
-    this.$vuetify.theme.dark = v;
-  }
-}
+});
 </script>