+++ src/piping-ui-auth.ts
@@ -1,38 +1,64 @@
-// NOTE: Use `const pipingUiAuthAsync = import(/* webpackChunkName: "piping-ui-auth" */ "@/_piping-ui-auth");` to import this file
-
 import {
   KeyExchangeParcel,
   keyExchangeParcelType,
   KeyExchangeV1Parcel,
   keyExchangeV1ParcelType,
-  Protection, VerificationStep,
+  Protection,
+  VerificationStep,
   VerifiedParcel,
   verifiedParcelType
 } from "@/datatypes";
-import {stringsByLang} from "@/strings/strings-by-lang";
 import type {Validation} from "io-ts";
-const utilsAsync = () => import("@/utils");
+import {sha256} from "@/utils/sha256";
+
+const openPgpUtilsAsync = () => import("@/utils/openpgp-utils");
 
 const jwkThumbprintAsync  = () => import("jwk-thumbprint");
 const uint8ArrayToStringAsync = () => import('binconv/dist/src/uint8ArrayToString').then(p => p.uint8ArrayToString);
+const stringToUint8ArrayAsync = () => import('binconv/dist/src/stringToUint8Array').then(p => p.stringToUint8Array);
 const urlJoinAsync = () => import('url-join').then(p => p.default);
 
 async function keyExchangePath(type: 'sender' | 'receiver', secretPath: string): Promise<string> {
-  const utils = await utilsAsync();
-  return utils.sha256(`${secretPath}/key_exchange/${type}`);
+  return await sha256(`${secretPath}/key_exchange/${type}`);
 }
 
-export async function verifiedPath(secretPath: string): Promise<string> {
-  const utils = await utilsAsync();
-  return utils.sha256(`${secretPath}/verified`);
+async function verifiedPath(secretPath: string): Promise<string> {
+  return await sha256(`${secretPath}/verified`);
 }
 
-export type KeyExchangeErrorCode = 'invalid_parcel_format' | 'invalid_v1_parcel_format' | 'different_key_exchange_version';
+export async function verify(serverUrl: string, secretPath: string, key: Uint8Array, verified: boolean, canceledPromise: Promise<void>) {
+  const openPgpUtils = await openPgpUtilsAsync();
+  const urlJoin = await urlJoinAsync();
+  const stringToUint8Array = await stringToUint8ArrayAsync();
+  const verifiedParcel: VerifiedParcel = {
+    verified,
+  };
+  const encryptedVerifiedParcel = await openPgpUtils.encrypt(
+    stringToUint8Array(JSON.stringify(verifiedParcel)),
+    key,
+  );
+  const path = urlJoin(serverUrl, await verifiedPath(secretPath));
+  const abortController = new AbortController();
+  canceledPromise.then(() => {
+    abortController.abort();
+  });
+  // Send verified or not
+  const res = await fetch(path, {
+    method: 'POST',
+    body: encryptedVerifiedParcel,
+    signal: abortController.signal,
+  });
+  // TODO: check res status
+  await res.text();
+}
+
+export type KeyExchangeErrorCode = 'send_failed' | 'receive_failed' | 'invalid_parcel_format' | 'invalid_v1_parcel_format' | 'different_key_exchange_version';
 type KeyExchangeResult =
   {type: "key", key: Uint8Array, verificationCode: string} |
-  {type: "error", errorCode: KeyExchangeErrorCode};
+  {type: "error", errorCode: KeyExchangeErrorCode} |
+  {type: "canceled"};
 
-export async function keyExchange(serverUrl: string, type: 'sender' | 'receiver', secretPath: string): Promise<KeyExchangeResult> {
+export async function keyExchange(serverUrl: string, type: 'sender' | 'receiver', secretPath: string, canceledPromise: Promise<void>): Promise<KeyExchangeResult> {
   const KEY_EXCHANGE_VERSION = 1;
   // 256 is max value for deriveBits()
   const KEY_BITS = 256;
@@ -55,11 +81,46 @@ export async function keyExchange(serverUrl: string, type: 'sender' | 'receiver'
   const urlJoin = await urlJoinAsync();
   const myPath = await keyExchangePath(type, secretPath);
   const peerPath = await keyExchangePath(type === 'sender' ? 'receiver' : 'sender', secretPath);
+  const abortController = new AbortController();
+  canceledPromise.then(() => {
+    abortController.abort();
+  });
   // Exchange
-  const [, peerRes] = await Promise.all([
-    fetch(urlJoin(serverUrl, myPath), {method: 'POST', body: JSON.stringify(keyExchangeParcel)}),
-    fetch(urlJoin(serverUrl, peerPath)),
-  ]);
+  const postResPromise = fetch(urlJoin(serverUrl, myPath), {
+    method: 'POST',
+    body: JSON.stringify(keyExchangeParcel),
+    signal: abortController.signal,
+  });
+  const peerResPromise = fetch(urlJoin(serverUrl, peerPath), {
+    signal: abortController.signal,
+  });
+  let postRes: Response;
+  try {
+    postRes = await postResPromise;
+  } catch (e: any) {
+    if (e.name === 'AbortError') {
+      return {type: "canceled"};
+    }
+    return {type: "error", errorCode: 'send_failed'};
+  }
+  if (postRes.status !== 200) {
+    return {type: "error", errorCode: 'send_failed'};
+  }
+  let peerRes: Response;
+  try {
+    [, peerRes] = await Promise.all([
+      postRes.text(),
+      peerResPromise,
+    ]);
+  } catch (e: any) {
+    if (e.name === 'AbortError') {
+      return {type: "canceled"};
+    }
+    return {type: "error", errorCode: 'receive_failed'};
+  }
+  if (peerRes.status !== 200) {
+    return {type: "error", errorCode: 'receive_failed'};
+  }
   const peerPublicKeyExchangeEither: Validation<KeyExchangeParcel> = keyExchangeParcelType.decode(JSON.parse(await peerRes.text()));
   if (peerPublicKeyExchangeEither._tag === 'Left') {
     return {type: "error", errorCode: 'invalid_parcel_format'};
@@ -99,15 +160,19 @@ async function generateVerificationCode(publicJwk1: JsonWebKey, publicJwk2: Json
     jwkThumbprintByEncoding(publicJwk1, 'SHA-256', 'hex'),
     jwkThumbprintByEncoding(publicJwk2, 'SHA-256', 'hex'),
   ];
-  const utils = await utilsAsync();
-  return (await utils.sha256(hashes.sort().join('-'))).substring(0, 32);
+  return (await sha256(hashes.sort().join('-'))).substring(0, 32);
 }
 
-export async function keyExchangeAndReceiveVerified(serverUrl: string, secretPath: string, protection: Protection, setVerificationStep: (step: VerificationStep) => void):
+type KeyExchangeAndReceiveVerifiedError =
+  { code: 'key_exchange_error', keyExchangeErrorCode: KeyExchangeErrorCode } |
+  { code: 'sender_not_verified' };
+
+export async function keyExchangeAndReceiveVerified(serverUrl: string, secretPath: string, protection: Protection, setVerificationStep: (step: VerificationStep) => void, canceledPromise: Promise<void>):
   Promise<
     {type: 'key', key: string | Uint8Array | undefined} |
-    {type: 'error', errorMessage: (lang: string) => string}
-    > {
+    {type: 'error', error: KeyExchangeAndReceiveVerifiedError } |
+    {type: 'canceled' }
+  > {
   switch (protection.type) {
     case 'raw':
       return {
@@ -121,14 +186,15 @@ export async function keyExchangeAndReceiveVerified(serverUrl: string, secretPat
       };
     case 'passwordless': {
       // Key exchange
-      const keyExchangeRes = await keyExchange(serverUrl, 'receiver', secretPath);
+      const keyExchangeRes = await keyExchange(serverUrl, 'receiver', secretPath, canceledPromise);
+      if (keyExchangeRes.type === 'canceled') {
+        return { type: "canceled" };
+      }
       if (keyExchangeRes.type === 'error') {
         setVerificationStep({type: 'error'});
         return {
           type: "error",
-          errorMessage(lang) {
-            return stringsByLang(lang)['key_exchange_error'](keyExchangeRes.errorCode);
-          }
+          error: { code: 'key_exchange_error', keyExchangeErrorCode: keyExchangeRes.errorCode },
         };
       }
       const {key, verificationCode} = keyExchangeRes;
@@ -136,9 +202,25 @@ export async function keyExchangeAndReceiveVerified(serverUrl: string, secretPat
       const uint8ArrayToString = await uint8ArrayToStringAsync();
       const urlJoin = await urlJoinAsync();
       const path = urlJoin(serverUrl, await verifiedPath(secretPath));
+      const abortController = new AbortController();
+      canceledPromise.then(() => {
+        abortController.abort();
+      });
       // Get verified or not
-      const res = await fetch(path);
-      const utils = await utilsAsync();
+      let res: Response;
+      try {
+        res = await fetch(path, {
+          signal: abortController.signal,
+        });
+      } catch (e: any) {
+        if (e.name === 'AbortError') {
+          return {type: "canceled"};
+        }
+        // TODO: return { type: "error" }
+        throw e;
+      }
+      // TODO: check res status
+      const utils = await openPgpUtilsAsync();
       // Decrypt body
       const decryptedBody: Uint8Array = await utils.decrypt(new Uint8Array(await res.arrayBuffer()), key);
       // Parse
@@ -146,9 +228,7 @@ export async function keyExchangeAndReceiveVerified(serverUrl: string, secretPat
       if (verifiedParcelEither._tag === "Left") {
         return {
           type: "error",
-          errorMessage(lang) {
-            return stringsByLang(lang)['key_exchange_error']('invalid_parcel_format');
-          }
+          error: { code: 'key_exchange_error', keyExchangeErrorCode: 'invalid_parcel_format'},
         };
       }
       const {verified} = verifiedParcelEither.right;
@@ -156,9 +236,7 @@ export async function keyExchangeAndReceiveVerified(serverUrl: string, secretPat
       if (!verified) {
         return {
           type: "error",
-          errorMessage(lang) {
-            return stringsByLang(lang)['sender_not_verified'];
-          }
+          error: { code: 'sender_not_verified' },
         };
       }
       return {