+++ vue.config.js
@@ -9,6 +9,10 @@ module.exports = {
   },
   configureWebpack: {
     plugins: [
+      new webpack.ProvidePlugin({
+        process: 'process',
+        Buffer: ['buffer', 'Buffer'],
+      }),
       // (base: https://medium.com/curofy-engineering/a-guide-to-inject-variable-into-your-code-using-webpack-36c49fcc1dcd)
       new webpack.DefinePlugin({
         PIPING_SERVER_URLS: process.env.PIPING_SERVER_URLS || JSON.stringify([
@@ -20,6 +24,12 @@ module.exports = {
         ]),
       })
     ],
+    resolve: {
+      fallback: {
+        process: require.resolve("process/browser"),
+        buffer: require.resolve("buffer/"),
+      },
+    },
   },
   // (from: https://github.com/vuetifyjs/vuetify/issues/8279#issuecomment-517900297)
   transpileDependencies: ['vuetify'],