+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.11.0] - 2022-09-20
+### Change
+- Improve error message in key exchange
+- Update dependencies
+- Improve cancellations especially for sending and viewing
+- (internal) Update @vue/cli-xxx
+- Improve UX by preloading components and libraries
+- Load icons asynchronously for performance
+- Update OpenPGP.js and reduce its built size for performance
+- Update theme color
+
 ## [0.10.0] - 2022-09-12
 ### Change
 - (internal) Use Composition API in the components
@@ -400,7 +411,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - First release
 
-[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.7.3...v0.8.0